initial version of how to use Translator() in __init__() of DimQuant
Actually, since Translator() depends on DimQuant, I cannot re-import
Translator into DimQuant. For this I have to derive what I temporarily
call TDimQuant from DimQuant, whose __init__() makes use of a
translator. The idea is to be able to register different translators
onto TDimQuant (as a classmethod), to use more than only SI units (e.g.
a user can define rules for imperical units and then register such a
translator). The implementation present in this commit is not as I want
it to be: TDimQuant should be called DimQuant, whereas DimQuant should
be something like BaseDimQuant. A user shouldn't have to interact with
BaseDimQuant and instead to only import
`>>> from dimensionalquantity import DimQuant as DQ`.
This renaming comes in a next commit.